% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{add_osm_tags}
\alias{add_osm_tags}
\alias{add_osm_features}
\title{Add multiple tags to an Overpass query as independent statements}
\usage{
add_osm_tags(opq, tags, bbox = NULL, key_exact = TRUE, value_exact = TRUE)

add_osm_features(
  opq,
  features,
  bbox = NULL,
  key_exact = TRUE,
  value_exact = TRUE
)
}
\arguments{
\item{opq}{An \code{overpass_query} object}

\item{tags}{A named list or vector with the format \code{list("<key>" = "<value>")} or \code{c("<key>" = "<value>")} or a character vector of
key-value pairs with keys and values enclosed in escape-formatted
quotations. See examples for details.}

\item{bbox}{optional bounding box for the query; must be set if no
opq query bbox has been set.}

\item{key_exact}{If FALSE, \code{key} is not interpreted exactly; see
\url{https://wiki.openstreetmap.org/wiki/Overpass_API}}

\item{value_exact}{If FALSE, \code{value} is not interpreted exactly}

\item{features}{Same as \code{tags}.}
}
\value{
An \link{opq} object.
}
\description{
Alternative version of \code{\link[=filter_osm_tags]{filter_osm_tags()}} for creating single queries
with multiple tags in independent statements. Key-value matching may be
controlled by using the filter symbols described in
\url{https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#By_tag_.28has-kv.29}.

\code{add_osm_features()} will be DEPRECATED in future versions.
}
\section{\code{filter_osm_tags} vs \code{add_osm_tags}}{

Tags defined within an \code{\link[=add_osm_tags]{add_osm_tags()}} call are combined with a
logical OR.

Chained calls to either \code{filter_osm_tags()} or \code{\link[=add_osm_tags]{add_osm_tags()}} combines
tags from these calls in a logical AND; this is analagous to chaining
\code{dplyr::filter()} on a data frame.

\code{add_osm_tags()} with only one tag is logically equivalent to
\code{filter_osm_tags()}.
}

\examples{
\dontrun{
q <- opq ("portsmouth usa") |>
    add_osm_tags (tags = list (
        "amenity" = "restaurant",
        "amenity" = "pub"
    ))

q <- opq ("portsmouth usa") |>
    add_osm_tags (tags = c (
        "\"amenity\"=\"restaurant\"",
        "\"amenity\"=\"pub\""
    ))
cat (opq_string (q))
# This extracts in a single query the same result as the following:
q1 <- opq ("portsmouth usa") |>
    filter_osm_tags (
        key = "amenity",
        value = "restaurant"
    )
q2 <- opq ("portsmouth usa") |>
    filter_osm_tags (key = "amenity", value = "pub")
c (osmdata_sf (q1), osmdata_sf (q2)) # all restaurants OR pubs

# Get objects with keys (`natural` OR `waterway`) AND `name`
q_keys <- opq ("Badia del Vallès", osm_types = "nwr", out = "tags") |>
    add_osm_tags (tags = list (natural = NULL, waterway = NULL)) |>
    filter_osm_tags (key = "name")
cat (opq_string (q_keys))
}
}
\references{
\url{https://wiki.openstreetmap.org/wiki/Map_Features}
}
\seealso{
\link{filter_osm_tags}

Other queries: 
\code{\link{bbox_to_string}()},
\code{\link{filter_osm_tags}()},
\code{\link{filter_osm_user}()},
\code{\link{getbb}()},
\code{\link{opq}()},
\code{\link{opq_around}()},
\code{\link{opq_csv}()},
\code{\link{opq_enclosing}()},
\code{\link{opq_osm_id}()},
\code{\link{opq_string}()},
\code{\link{overpass_status}()}
}
\concept{queries}
