% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leaflet.r
\name{addOverpass}
\alias{addOverpass}
\title{Add an Overpass Object to a Leaflet Plot}
\usage{
addOverpass(overpass_obj)
}
\arguments{
\item{overpass_obj}{the result of a call to \code{\link{overpass_query}}}

\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}}}
}
\value{
\code{\link[leaflet]{leaflet}} map object
}
\description{
Add an Overpass Object to a Leaflet Plot
}

