% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbb.R
\name{area_to_string}
\alias{area_to_string}
\title{Convert a data.frame with OSM types and ids to a string}
\usage{
area_to_string(area)
}
\arguments{
\item{area}{\link{data.frame} with \code{osm_type} and \code{osm_id} columns.}
}
\value{
A character string representing a set of OSM objects in overpass query
language. For example: \code{"relation(id:11747082)"} represents the area of
the Catalan Countries. A set of objects can also be represented for multirow
\code{area} (e.g. "relation(id:11747082,307833); way(id:22422490)")
}
\description{
This function converts areas defined by type and id into a string for use in
opq to delimitate the search area
}
\examples{
\dontrun{
area_to_string(getbb("Catalan Countries", format_out = "data.frame"))
}
}
\seealso{
Other queries: 
\code{\link{add_osm_features}()},
\code{\link{add_osm_feature}()},
\code{\link{bbox_to_string}()},
\code{\link{getarea}()},
\code{\link{getbb}()},
\code{\link{opq_around}()},
\code{\link{opq_enclosing}()},
\code{\link{opq_osm_id}()},
\code{\link{opq_string}()},
\code{\link{opq}()},
\code{\link{overpass_status}()}
}
\concept{queries}
