% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{available_tags}
\alias{available_tags}
\title{List tags associated with a feature}
\usage{
available_tags(feature, taginfo = FALSE)
}
\arguments{
\item{feature}{feature to retrieve}

\item{taginfo}{If TRUE retrieve the full list of OSM tags for \code{feature}.}
}
\value{
Character vector of all known tags for a feature.
}
\description{
List most commonly used tags asociated to a \code{feature} as described in OSM's
\href{https://wiki.openstreetmap.org/wiki/Map_Features}{Map Features page}.
Optionally return the full list of tags used for \code{feature} via the
\href{https://taginfo.openstreetmap.org/}{\code{taginfo} API}.
}
\note{
requires internet access
}
\examples{
\dontrun{
available_tags("aerialway")
}
}
\references{
\url{https://wiki.openstreetmap.org/wiki/Map_Features}
}
