% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbox_to_string}
\alias{bbox_to_string}
\title{Convert a named matrix or a named vector (or an unnamed vector) return a string}
\usage{
bbox_to_string(bbox)
}
\arguments{
\item{bbox}{bounding box}
}
\description{
This function converts a bounding box into a string for use in web apis
}

