% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{filter_osm_user}
\alias{filter_osm_user}
\title{Add an user filter to an Overpass query}
\usage{
filter_osm_user(opq, user, touched = FALSE, is_uid)
}
\arguments{
\item{opq}{An \code{overpass_query} object.}

\item{user}{A vector of user names or user ids. If \code{is.numeric(user)} or
\verb{all(grepl(^[0-9]+$, user})), assumes that \code{user} is a user id.}

\item{touched}{If \code{TRUE}, selects objects of which at least one version has been edited
by \code{user}. If \code{FALSE} (default), selects objects with \code{user} as the last editor
only.}

\item{is_uid}{If \code{FALSE}, assume that \code{user} are user names. Useful to filter for a
user name which contains numbers only and otherwise would be considered an user id.}
}
\value{
An \link{opq} object.
}
\description{
Select objects last edited or ever touched by a specific user.
}
\details{
The order of \code{filter_osm_user()} in a query construction is not relevant and is applied
to the whole query (i.e. it affects all the statements of the query).

See
\url{https://dev.overpass-api.de/overpass-doc/en/criteria/misc_criteria.html#by_user}
for details.
}
\examples{
# Notice the "::user" and "::uid" fields in the oqp_csv()
q_csv <- opq (bbox = "relation(id:11755232)", out = "meta", osm_type = "node") |>
    filter_osm_user (user = "jmaspons") |>
    add_osm_feature (key = "name") |>
    opq_csv (c ("::type", "::id", "name", "name:ca", "::user", "::uid"))
cat (opq_string (q_csv))
# Warning: csv queries can fail without errors in long running queries. For timeouts,
# it returns empty results.
\dontrun{
d_csv <- osmdata_data_frame (q_csv)
d_csv
}

q_touched <- opq (
    bbox = "relation(id:11755232)", out = "meta",
    osm_type = "node", timeout = 100
) |>
    filter_osm_user (user = "jmaspons", touched = TRUE) |>
    add_osm_feature (key = "name")
cat (opq_string (q_touched))
\dontrun{
d_touched <- osmdata_data_frame (q_touched)
d_touched
}
}
\seealso{
Other queries: 
\code{\link{add_osm_tags}()},
\code{\link{bbox_to_string}()},
\code{\link{filter_osm_tags}()},
\code{\link{getbb}()},
\code{\link{opq}()},
\code{\link{opq_around}()},
\code{\link{opq_csv}()},
\code{\link{opq_enclosing}()},
\code{\link{opq_osm_id}()},
\code{\link{opq_string}()},
\code{\link{overpass_status}()}
}
\concept{queries}
