% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-highways.R
\name{get_highways}
\alias{get_highways}
\title{get_highways}
\usage{
get_highways(bbox = NULL)
}
\arguments{
\item{bbox}{the bounding box within which highways should be downloaded.  A
2-by-2 matrix of 4 elements with columns of min and max values, and rows of x
and y values.}
}
\value{
A 'SpatialLinesDataFrame' object containing all the highways within
the given bounding box.
}
\description{
Extracts highways from an overpass download.
}

