% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{get_nodes}
\usage{
get_nodes(bbox, key, url_download, raw_data = FALSE, verbose = FALSE)
}
\arguments{
\item{bbox}{the bounding box within which ways should be downloaded.  A
2-by-2 matrix of 4 elements with columns of min and max values, and rows of x
and y values.}

\item{key}{The OpenStreetMap key to be passed to the overpass API query, or
to be extracted from pre-downloaded data passed as \code{url_download}}

\item{url_download}{Data as directly downloaded from the overpass API and
returned with \code{raw_data=TRUE}. This may be subsequently passed to
\code{get_ways} in order to extract particular \code{key-value} combinations}

\item{raw_data}{If TRUE, \code{get_ways} returns unprocessed data as directly
returned from the overpass API query.}

\item{verbose}{If TRUE, provides notification of progress}
}
\value{
A \code{SpatialPointsDataFrame} object containing all the nodes within
the given bounding box.
}
\description{
Extracts nodes from an overpass download.
}

