% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ways.R
\name{get_ways}
\alias{get_ways}
\title{get_ways}
\usage{
get_ways(bbox, key, value, extra_pairs, raw_data = FALSE, verbose = FALSE)
}
\arguments{
\item{bbox}{the bounding box within which ways should be downloaded.  A
2-by-2 matrix of 4 elements with columns of min and max values, and rows of x
and y values.}

\item{key}{The OpenStreetMap key to be passed to the overpass API query, or
to be extracted from pre-downloaded data passed as \code{url_download}}

\item{value}{OSM value to match to key. If \code{NULL}, all keys will be
returned.  Negation is specified by \code{!value}.}

\item{extra_pairs}{A list of additional \code{key-value} pairs to be passed
to the overpass API.}

\item{raw_data}{If TRUE, \code{get_ways} returns unprocessed data as directly
returned from the overpass API query.}

\item{verbose}{If TRUE, provides notification of progress}
}
\value{
A \code{SpatialLinesDataFrame} object containing all the ways within
the given bounding box.
}
\description{
Extracts ways from an overpass download.
}

