% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects2.R
\name{get_xml2_doc}
\alias{get_xml2_doc}
\title{get_xml2_doc}
\usage{
get_xml2_doc(bbox = NULL)
}
\arguments{
\item{bbox}{the bounding box within which highways should be downloaded.  A
2-by-2 matrix of 4 elements with columns of min and max values, and rows of x
and y values.}
}
\value{
An xml2 'xml_document'
}
\description{
Adapted from 'https://github.com/hrbrmstr/overpass/' to extract an overpass
API query using 'xml2'.  Only implemented here for highways.
}

