% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{get_xml_doc3}
\alias{get_xml_doc3}
\title{get_xml_doc3}
\usage{
get_xml_doc3(bbox = NULL)
}
\arguments{
\item{bbox}{the bounding box within which highways should be downloaded.  A
2-by-2 matrix of 4 elements with columns of min and max values, and rows of x
and y values.}
}
\value{
An XML 'XMLAbstractDocument'
}
\description{
Stripped from 'osmplotr/extract_osm_objects' to extract highways only for
timing comparison.
}

