% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbb.R
\name{getbb}
\alias{getbb}
\title{Get bounding box for a given place name}
\usage{
getbb(place_name, viewbox = NULL, format_out = c("matrix", "string"),
  base_url = "https://nominatim.openstreetmap.org",
  featuretype = "settlement", silent = TRUE)
}
\arguments{
\item{place_name}{The name of the place you're searching for}

\item{viewbox}{The bounds in which you're searching}

\item{format_out}{Character string indicating output format: matrix (default - see \code{\link{bbox}})
or string (see \code{\link{bbox_to_string}})}

\item{base_url}{Base website from where data is queried}

\item{featuretype}{The type of OSM feature (settlement is default)}

\item{silent}{Should the API be printed to screen? FALSE by default}
}
\description{
Uses online services to convert a text string into a bounding box
}
\examples{
if(curl::has_internet()){
  getbb("Salzburg")
  # refine the search to the USA
  place_name = "Hereford"
  getbb(place_name)
  bb_usa = getbb("United States")
  viewbox = bbox_to_string(bb_usa)
  getbb(place_name, viewbox, silent = FALSE) # not working
}

}

