% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{list_overpass_urls}
\alias{list_overpass_urls}
\title{list_overpass_urls}
\usage{
list_overpass_urls()
}
\value{
A character vector of Overpass API interpreter URLs.
}
\description{
List public Overpass API URLs. These are mirrors sampled by default
when the package loads.
}
\details{
For further details, see
\url{https://wiki.openstreetmap.org/wiki/Overpass_API#Public_Overpass_API_instances}
and \url{https://github.com/ropensci/osmdata/pull/149}.
}
\examples{
list_overpass_urls ()
}
\seealso{
Other overpass:
\code{\link[=get_overpass_url]{get_overpass_url()}},
\code{\link[=set_overpass_url]{set_overpass_url()}}
}
\concept{overpass}
