% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{opq}
\alias{opq}
\title{Build an Overpass query}
\usage{
opq(bbox = NULL)
}
\arguments{
\item{bbox}{base bounding box to use with the features. Must set the individual
feature bbox values if this value is not set. Can be a matrix (i.e. what
\code{sp::bbox} returns), a string with values ("left,bottom,top,right"),
a vector of length 4, or (if tmap is installed) a character place name.
If the vector is named, the names will be used,
otherwise, you should ensure the vector is in \code{c(top, left, bottom, right)}
order.}
}
\value{
\code{opq} object
}
\description{
Build an Overpass query
}
\examples{
\dontrun{
q <- opq(bbox=c(-70.8229993, 43.0135509, -70.7280563, 43.0996118)) \%>\%
     add_feature("amenity", "pub", ) \%>\%
     add_feature("amenity", "restaurant") \%>\%
     add_feature("amenity", "library")
q
overpass_query(q) -> reading_noms
sp::plot(reading_noms$osm_nodes)
p <- overpass_query(opq("Leeds") \%>\%
  add_feature("leisure", "park"))
sp::plot(p$osm_polygons[1,])
}
}

