% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{opq_string}
\alias{opq_string}
\alias{opq_to_string}
\title{Convert an overpass query into a text string}
\usage{
opq_string(opq)
}
\arguments{
\item{opq}{An \code{overpass_query} object}
}
\value{
Character string to be submitted to the overpass API
}
\description{
Convert an osmdata query of class opq to a character string query to
be submitted to the overpass API.
}
\examples{
\dontrun{
q <- opq ("hampi india")
opq_string (q)
}
}
\seealso{
Other queries: 
\code{\link{add_osm_features}()},
\code{\link{add_osm_feature}()},
\code{\link{bbox_to_string}()},
\code{\link{getbb}()},
\code{\link{opq_around}()},
\code{\link{opq_csv}()},
\code{\link{opq_enclosing}()},
\code{\link{opq_osm_id}()},
\code{\link{opq}()},
\code{\link{overpass_status}()}
}
\concept{queries}
