% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{opq_to_string}
\alias{opq_to_string}
\title{Convert an osmdata query of class \code{opq} to a character string query to
be submitted to the overpass API}
\usage{
opq_to_string(opq)
}
\arguments{
\item{opq}{An \code{overpass_query} object}
}
\value{
Character string to be submitted to the overpass API
}
\description{
Convert an osmdata query of class \code{opq} to a character string query to
be submitted to the overpass API
}
\examples{
\dontrun{
q <- opq ("hampi india")
opq_to_string (q)
}
}
