% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_extract.R
\name{osm_lines}
\alias{osm_lines}
\title{Extract all \code{osm_lines} from an osmdata object}
\usage{
osm_lines(dat, id)
}
\arguments{
\item{dat}{An object of class \code{osmdata}}

\item{id}{OMS identification of object for which lines are to be extracted}
}
\value{
An \code{sf} Simple Features Collection of linestrings
}
\description{
If \code{id} is of a point object, \code{osm_lines} will return all lines
containing that point. If \code{id} is of a line object, \code{osm_lines}
will return all lines which intersect the given line.
}
