% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm-extract.R
\name{osm_multilines}
\alias{osm_multilines}
\title{Extract all \code{osm_multilines} from an osmdata object}
\usage{
osm_multilines(dat, id)
}
\arguments{
\item{dat}{An object of class \code{osmdata}}

\item{id}{OSM identification of one of more objects for which multilines are
to be extracted}
}
\value{
An \code{sf} Simple Features Collection of multilines
}
\description{
\code{id} must be of an \code{osm_point} or \code{osm_line} object (and can
not be the \code{id} of an \code{osm_polygon} object because multilines by
definition contain no polygons.  \code{osm_multilines} returns any multiline
object(s) which contain the object specified by \code{id}.
}
