% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm-extract.R
\name{osm_points}
\alias{osm_points}
\title{Extract all \code{osm_points} from an osmdata object}
\usage{
osm_points(dat, id)
}
\arguments{
\item{dat}{An object of class \code{osmdata}}

\item{id}{OSM identification of one or more objects for which points are to
be extracted}
}
\value{
An \code{sf} Simple Features Collection of points
}
\description{
Extract all \code{osm_points} from an osmdata object
}
