% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{osmdata-deprecated}
\alias{osmdata-deprecated}
\title{Deprecated functions and arguments in osmdata}
\description{
These functions and arguments have been deprecated and will be removed in a future release.
}
\section{Deprecated arguments}{

\describe{
\item{\code{nodes_only = TRUE} (in \code{\link[=opq]{opq()}})}{
This argument has been replaced by \code{osm_types = "node"}.
Since version 0.3, using \code{nodes_only} will produce a deprecation warning.
}
}
}

\section{Deprecated functions}{

\describe{
\item{\code{osmdata_sp()}}{
Please use \code{\link[=osmdata_sf]{osmdata_sf()}} or \code{\link[=osmdata_sc]{osmdata_sc()}} instead.
Since version 0.3, using \code{\link[=osmdata_sp]{osmdata_sp()}} will produce a deprecation warning.
}
}
}

\keyword{internal}
