% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata-package.R
\name{osmdata-exports}
\alias{\%>\%}
\alias{osmdata-exports}
\title{osmdata exported operators}
\description{
The following functions are imported and then re-exported
from the osmdata package to enable use of the magrittr/dplyr
pipe operator with no additional library calls
}

