% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata-package.R
\docType{package}
\name{osmdata-package}
\alias{osmdata-package}
\title{Import OpenStreetMap data in 'sf', 'SC', 'sp', 'data.frame' and 'xml' formats}
\description{
Imports OpenStreetMap (OSM) data into R as 'sf', 'SC', 'sp', 'data.frame' or
'xml_document' objects. OSM data are extracted from the overpass API and
processed with very fast C++ routines for return to R. The package enables
simple overpass queries to be constructed without the user necessarily
understanding the syntax of the overpass query language, while retaining the
ability to handle arbitrarily complex queries. Functions are also provided to
enable recursive searching between different kinds of OSM data (for example,
to find all lines which intersect a given point).
}
\section{Functions to Prepare Queries}{

\itemize{
\item \code{\link[=getbb]{getbb()}}: Get bounding box for a given place name
\item \code{\link[=bbox_to_string]{bbox_to_string()}}: Convert a named matrix or a named vector (or an
unnamed vector) return a string
\item \code{\link[=overpass_status]{overpass_status()}}: Retrieve status of the overpass API
\item \code{\link[=opq]{opq()}}: Build an overpass query
\item \code{\link[=filter_osm_tags]{filter_osm_tags()}}: Add tags to all current statements of an Overpass query
\item \code{\link[=add_osm_tags]{add_osm_tags()}}: Add multiple tags to independent statments of an Overpass query
\item \code{\link[=filter_osm_user]{filter_osm_user()}}: Add an user filter to an Overpass query
\item \code{\link[=opq_string]{opq_string()}}: Convert an osmdata query to overpass API string
}
}

\section{Functions to Get Additional OSM Information}{

\itemize{
\item \code{\link[=available_features]{available_features()}}: List recognised features in OSM
\item \code{\link[=available_tags]{available_tags()}}: List tags associated with a feature
}
}

\section{Functions to Extract OSM Data}{

\itemize{
\item \code{\link[=osmdata_data_frame]{osmdata_data_frame()}}: Return OSM data in \code{\link{data.frame}} format
\item \code{\link[=osmdata_sc]{osmdata_sc()}}: Return OSM data in \pkg{silicate} format
\item \code{\link[=osmdata_sf]{osmdata_sf()}}: Return OSM data in \pkg{sf} format
\item \code{\link[=osmdata_sp]{osmdata_sp()}}: Return OSM data in \pkg{sp} format (DEPRECATED)
\item \code{\link[=osmdata_xml]{osmdata_xml()}}: Return OSM data in \pkg{xml2} format
}
}

\section{Functions to Search Data}{

\itemize{
\item \code{\link[=osm_points]{osm_points()}}: Extract all \code{osm_points} objects
\item \code{\link[=osm_lines]{osm_lines()}}: Extract all \code{osm_lines} objects
\item \code{\link[=osm_polygons]{osm_polygons()}}: Extract all \code{osm_polygons} objects
\item \code{\link[=osm_multilines]{osm_multilines()}}: Extract all \code{osm_multilines} objects
\item \code{\link[=osm_multipolygons]{osm_multipolygons()}}: Extract all \code{osm_multipolygons} objects
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/osmdata/}
  \item \url{https://github.com/ropensci/osmdata}
  \item Report bugs at \url{https://github.com/ropensci/osmdata/issues}
}

}
\author{
Joan Maspons, Mark Padgham, Bob Rudis, Robin Lovelace, Maëlle Salmon
}
\concept{package}
\keyword{internal}
