\name{osmdatar-package}
\alias{osmdatar-package}
\alias{osmdatar}
\docType{package}
\title{
\packageTitle{osmdatar}
}
\description{
\packageDescription{osmdatar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{osmdatar}
\packageIndices{osmdatar}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{osmdatar}

Maintainer: \packageMaintainer{osmdatar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
