% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdatar.R
\docType{package}
\name{osmdatar}
\alias{osmdatar}
\alias{osmdatar-package}
\title{osmdata}
\description{
Very fast download of OpenStreetMap (OSM) data into 'sp' objects. OSM data
are extracted from the overpass API and processed with very fast C++ routines
for return to R.
}
\section{Functions}{

\tabular{ll}{
'get_nodes'\tab Extract nodes from OSM data and return as
SpatialPointsDataFrame\cr
'get_ways'\tab Extract ways from OSM data and return as
SpatialLinesDataFrame\cr
'get_polygons'\tab Extract polygons from OSM data and return as
SpatialPolygonsDataFrame\cr
}
}

