% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overpass-package.r
\name{overpass-exports}
\alias{\%>\%}
\alias{overpass-exports}
\title{overpass exported operators}
\description{
The following functions are imported and then re-exported
from the overpass package to enable use of the magrittr/dplyr
pipe operator with no additional library calls
}

