% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overpass-package.r
\docType{package}
\name{overpass}
\alias{overpass}
\alias{overpass-package}
\title{Tools to Work With the OpenStreetMap (OSM) Overpass API}
\description{
Tools to Work With the OpenStreetMap (OSM) Overpass API
}
\author{
Bob Rudis, Robin Lovelace
}

