% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overpass_query.r
\name{overpass_query}
\alias{overpass_query}
\title{Issue OSM Overpass Query}
\usage{
overpass_query(query)
}
\arguments{
\item{query}{OSM Overpass query. Please note that the function is in ALPHA
dev stage and needs YOU to specify that the output type is XML.
However, you can use Overpass XML or Overpass QL formats.}
}
\value{
If the \code{query} result only has OSM \code{node}s then the function
        will return a \code{SpatialPointsDataFrame} with the \code{node}s.\cr\cr
        If the \code{query} result has OSM \code{way}s then the function
        will return a \code{SpatialLinesDataFrame} with the \code{way}s\cr\cr
        \code{relations}s are not handled yet.
}
\description{
Issue OSM Overpass Query
}
\note{
wrap function with \code{httr::with_verbose} if you want to see the \code{httr}
      query (useful for debugging connection issues)
}

