% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overpass-query.R
\name{overpass_status}
\alias{overpass_status}
\title{Retrieve status of the Overpass API}
\usage{
overpass_status(quiet = FALSE)
}
\arguments{
\item{quiet}{if \code{FALSE} display a status message}
}
\value{
an invisible list of whether the API is available along with the
text of the message from Overpass and the timestamp of the
next available slot
}
\description{
Retrieve status of the Overpass API
}
\seealso{
Other queries: 
\code{\link{add_osm_features}()},
\code{\link{add_osm_feature}()},
\code{\link{bbox_to_string}()},
\code{\link{getbb}()},
\code{\link{opq_around}()},
\code{\link{opq_csv}()},
\code{\link{opq_enclosing}()},
\code{\link{opq_osm_id}()},
\code{\link{opq_string}()},
\code{\link{opq}()}
}
\concept{queries}
