% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects2.R
\name{process_xml2_doc}
\alias{process_xml2_doc}
\title{process_xml2_doc}
\usage{
process_xml2_doc(doc)
}
\arguments{
\item{doc}{An XML document extracted with 'get_xml_doc'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Implements code from 'https://github.com/hrbrmstr/overpass/' to extract
highways from an XML doc extracted with 'get_xml_doc'
}
\examples{
bbox <- matrix (c (-0.13, 51.5, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml2_doc (bbox=bbox)
obj <- process_xml2_doc2 (doc)
}

