% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects.R
\name{process_xml_doc}
\alias{process_xml_doc}
\title{process_xml_doc}
\usage{
process_xml_doc(doc)
}
\arguments{
\item{doc}{An XML document extracted with 'get_xml_doc'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Adapted from 'osmplotr/extract_osm_objects' to extract SpatialLines from an
XML doc extracted with 'get_xml_doc'
}
\examples{
bbox <- matrix (c (-0.13, 51.5, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc (bbox=bbox)
obj <- process_xml_doc (doc)
}

