% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{process_xml_doc3}
\alias{process_xml_doc3}
\title{process_xml_doc3}
\usage{
process_xml_doc3(txt)
}
\arguments{
\item{doc}{A text document extracted with 'get_xml_doc3'}
}
\value{
A SpatialLinesDataFrame
}
\description{
process_xml_doc3
}
\examples{
bbox <- matrix (c (-0.13, 51.5, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc (bbox=bbox)
obj <- process_xml_doc (doc)
}

