% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{process_xml_doc3a}
\alias{process_xml_doc3a}
\title{process_xml_doc3a}
\usage{
process_xml_doc3a(txt)
}
\arguments{
\item{doc}{A text document extracted with 'get_xml_doc3'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Uses an explicit loop to convert list objects returned from 'get_highways'
into Spatial Lines objects
}
\examples{
bbox <- matrix (c (-0.13, 51.5, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc3 (bbox=bbox)
obj <- process_xml_doc3a (doc)
}

