% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{process_xml_doc3b}
\alias{process_xml_doc3b}
\title{process_xml_doc3b}
\usage{
process_xml_doc3b(txt)
}
\arguments{
\item{doc}{A text document extracted with 'get_xml_doc3'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Differs from 'process_xml_doc3a' only in calling the 'Rcpp' function
'get_highways_with_id' instead of 'get_highways'. The list elements returned
from the latter each have 2 columns (lon, lat), while they former prepends a
third column with the ID of the OSM highway.
}
\examples{
bbox <- matrix (c (-0.12, 51.52, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc3 (bbox=bbox)
obj <- process_xml_doc3b (doc)
}

