% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{process_xml_doc3c}
\alias{process_xml_doc3c}
\title{process_xml_doc3c}
\usage{
process_xml_doc3c(txt)
}
\arguments{
\item{doc}{A text document extracted with 'get_xml_doc3'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Differs from 'process_xml_doc3b' only in replacing the explicit loop with an
'lapply'.
}
\examples{
bbox <- matrix (c (-0.12, 51.52, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc3 (bbox=bbox)
obj <- process_xml_doc3c (doc)
}

