% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{process_xml_doc3d}
\alias{process_xml_doc3d}
\title{process_xml_doc3d}
\usage{
process_xml_doc3d(txt)
}
\arguments{
\item{doc}{A text document extracted with 'get_xml_doc3'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Instead of the loops or 'lapply' calls of 'process_xml_doc3b/c', this version
uses the 'dplyr::group_by' approach of hrbrmstr. This is definitely slower
than the above versions!
}
\examples{
bbox <- matrix (c (-0.12, 51.52, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc3 (bbox=bbox)
obj <- process_xml_doc3d (doc)
}

