% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{process_xml_doc3e}
\alias{process_xml_doc3e}
\title{process_xml_doc3e}
\usage{
process_xml_doc3e(txt)
}
\arguments{
\item{doc}{A text document extracted with 'get_xml_doc3'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Uses the Rcpp function 'get_highways_sp' which constructs S4 sp objects
within Rcpp. This '_sp' function constructs 'SpatialLines', while
'process_xml_docf' uses the equivalent function that constructs the entire
'SpatialLinesDataFrame' within Rcpp
}
\examples{
bbox <- matrix (c (-0.13, 51.5, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc3 (bbox=bbox)
obj <- process_xml_doc3e (doc)
}

