% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-osm-objects3.R
\name{process_xml_doc3f}
\alias{process_xml_doc3f}
\title{process_xml_doc3f}
\usage{
process_xml_doc3f(txt)
}
\arguments{
\item{doc}{A text document extracted with 'get_xml_doc3'}
}
\value{
A SpatialLinesDataFrame
}
\description{
Uses the Rcpp function 'get_highways_spLines' which constructs the entire
SpatialLinesDataFrame object within Rcpp.
}
\examples{
bbox <- matrix (c (-0.13, 51.5, -0.11, 51.52), nrow=2, ncol=2)
doc <- get_xml_doc3 (bbox=bbox)
obj <- process_xml_doc3f (doc)
}

