% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{qry_to_string}
\alias{qry_to_string}
\title{Convert an osmdata query of class \code{opq} to a character string query to
be submitted to the overpass API}
\usage{
qry_to_string(qry)
}
\arguments{
\item{qry}{Overpass query object}
}
\value{
Character string to be submitted to the overpass API
}
\description{
Convert an osmdata query of class \code{opq} to a character string query to
be submitted to the overpass API
}
\note{
The final query can be obtained from 
\code{paste0 (c (query$features, query$suffix), collapse="\n")}
}
