% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_nodes}
\alias{rcpp_get_nodes}
\title{rcpp_get_nodes}
\usage{
rcpp_get_nodes(st)
}
\arguments{
\item{st}{Text contents of an overpass API query}
}
\value{
A SpatialPointsDataFrame contains all nodes and associated data
}
\description{
Extracts all nodes from an overpass API query
}

