% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_points}
\alias{rcpp_get_points}
\title{rcpp_get_points}
\usage{
rcpp_get_points(st)
}
\arguments{
\item{st}{Text contents of an overpass API query}
}
\value{
A \code{SpatialPointsDataFrame} contains all nodes and associated data
}
\description{
Extracts all nodes from an overpass API query
}

