% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_polygons}
\alias{rcpp_get_polygons}
\title{rcpp_get_polygons}
\usage{
rcpp_get_polygons(st)
}
\arguments{
\item{st}{Text contents of an overpass API query}
}
\value{
A SpatialLinesDataFrame contains all polygons and associated data
}
\description{
Extracts all polygons from an overpass API query
}

