% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_get_ways}
\alias{rcpp_get_ways}
\title{rcpp_get_ways}
\usage{
rcpp_get_ways(st)
}
\arguments{
\item{st}{Text contents of an overpass API query}
}
\value{
A \code{SpatialLinesDataFrame} contains all ways and associated data
}
\description{
Extracts all ways from an overpass API query
}

