% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_osmdata}
\alias{rcpp_osmdata}
\title{rcpp_osmdata}
\usage{
rcpp_osmdata(st)
}
\arguments{
\item{st}{Text contents of an overpass API query}
}
\value{
Rcpp::List objects of OSM data
}
\description{
Extracts all polygons from an overpass API query
}

