% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_osm.R
\name{read_osm}
\alias{read_osm}
\title{Read OSM Overpass Response}
\usage{
read_osm(osm_response)
}
\arguments{
\item{osm_response}{file containing an OSM Overpass XML response}
}
\value{
If the \code{query} result only has OSM \code{node}s then the function
        will return a \code{SpatialPointsDataFrame} with the \code{node}s.\cr\cr
        If the \code{query} result has OSM \code{way}s then the function
        will return a \code{SpatialLinesDataFrame} with the \code{way}s\cr\cr
        \code{relations}s are not handled yet.\cr\cr
}
\description{
Read OSM Overpass Response
}

