% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_overpass_url}
\alias{set_overpass_url}
\title{set_overpass_url}
\usage{
set_overpass_url(overpass_url)
}
\arguments{
\item{overpass_url}{The desired overpass API URL}
}
\value{
The overpass API URL
}
\description{
Set the URL of the specified overpass API. Possible APIs with global
coverage are returned by \code{\link[=list_overpass_urls]{list_overpass_urls()}}.
}
\details{
For further details, see
\url{https://wiki.openstreetmap.org/wiki/Overpass_API}
}
\examples{
\dontrun{
set_overpass_url (list_overpass_urls () [1])
}
}
\seealso{
Other overpass:
\code{\link[=get_overpass_url]{get_overpass_url()}},
\code{\link[=list_overpass_urls]{list_overpass_urls()}}
}
\concept{overpass}
