% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbb.R
\name{url_available}
\alias{url_available}
\title{Check whether an API URL is available or not}
\usage{
url_available(base_url, max_trials = 10)
}
\arguments{
\item{base_url}{The URL to be checked}

\item{max_trials}{Maximum number of attempts before returning FALSE}
}
\value{
TRUE if URL available, otherwise FALSE
}
\description{
Check whether an API URL is available or not
}
