% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{add_feature}
\alias{add_feature}
\title{Add a feature to an Overpass query}
\usage{
add_feature(opq, key, value, exact = TRUE, bbox = NULL)
}
\arguments{
\item{opq}{An \code{overpass_query} object}

\item{key}{feature key}

\item{value}{value for feature key}

\item{exact}{If FALSE, \code{value} is not interpreted exactly; see
\url{http://wiki.openstreetmap.org/wiki/Overpass_API/Language_Guide#Non-exact_names}}

\item{bbox}{optional bounding box for the feature query; must be set if no
opq query bbox has been set}
}
\value{
\code{opq} object
}
\description{
Add a feature to an Overpass query
}
\note{
The actual query submitted to the overpass API can be obtained from
\link{opq_to_string}
}
\examples{
\dontrun{
q <- getbb ("portsmouth", display_name_contains="United States") \%>\% opq () \%>\% 
        add_feature("amenity", "restaurant") \%>\%
        add_feature("amenity", "pub") 
osmdata_sf (q) # all objects that are restaurants AND pubs (there are none!)
q1 <- getbb ("portsmouth", display_name_contains="United States") \%>\% opq () \%>\% 
        add_feature("amenity", "restaurant") 
q2 <- getbb ("portsmouth", display_name_contains="United States") \%>\% opq () \%>\% 
        add_feature("amenity", "pub") 
c (osmdata_sf (q1), osmdata_sf (q1)) # all objects that are restaurants OR pubs
}
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Map_Features}
}
