% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.r
\name{available_features}
\alias{available_features}
\title{List recognized features in OSM Overpass}
\usage{
available_features()
}
\value{
character vector of all known features
}
\description{
List recognized features in OSM Overpass
}
\note{
requires internet access
}
\examples{
available_features()
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Map_Features}
}

