% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{available_features}
\alias{available_features}
\title{List recognized features in OSM}
\usage{
available_features()
}
\value{
character vector of all known features
}
\description{
List recognized features in OSM
}
\note{
requires internet access
}
\examples{
\dontrun{
available_features()
}
}
\references{
\url{https://wiki.openstreetmap.org/wiki/Map_Features}
}
