% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features.r
\name{available_tags}
\alias{available_tags}
\title{List tags associated with a feature}
\usage{
available_tags(feature)
}
\value{
character vector of all known tags for a feature
}
\description{
List tags associated with a feature
}
\note{
requires internet access
}
\examples{
available_tags("aerialway")
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Map_Features}
}

