% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{available_tags}
\alias{available_tags}
\title{List tags associated with a feature}
\usage{
available_tags(feature)
}
\arguments{
\item{feature}{feature to retrieve}
}
\value{
character vector of all known tags for a feature
}
\description{
List tags associated with a feature
}
\note{
requires internet access
}
\examples{
\dontrun{
available_tags("aerialway")
}
}
\references{
\url{https://wiki.openstreetmap.org/wiki/Map_Features}
}
