% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbb.R
\name{bbox_to_string}
\alias{bbox_to_string}
\title{Convert a named matrix or a named vector (or an unnamed vector) return a string}
\usage{
bbox_to_string(bbox)
}
\arguments{
\item{bbox}{bounding box as character, matrix or vector. If character,
numeric bbox will be extracted with \link{getbb} Unnamed vectors will be
sorted appropriately and must merely be in the order (x, y, x, y).}
}
\description{
This function converts a bounding box into a string for use in web apis
}
\examples{
\dontrun{
bbox_to_string (getbb ("hampi india"))
}
}
