% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_overpass_url}
\alias{get_overpass_url}
\title{get_overpass_url}
\usage{
get_overpass_url()
}
\value{
The overpass API URL
}
\description{
Return the URL of the specified overpass API. Default is
\url{https://overpass-api.de/api/interpreter}.
}
\seealso{
\code{\link[=set_overpass_url]{set_overpass_url()}}
}
