% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ways.R
\name{get_ways}
\alias{get_ways}
\title{get_ways}
\usage{
get_ways(bbox = NULL)
}
\arguments{
\item{bbox}{the bounding box within which ways should be downloaded.  A
2-by-2 matrix of 4 elements with columns of min and max values, and rows of x
and y values.}
}
\value{
A 'SpatialLinesDataFrame' object containing all the ways within
the given bounding box.
}
\description{
Extracts ways from an overpass download.
}

