% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opq.R
\name{opq}
\alias{opq}
\title{Begin building an Overpass query}
\usage{
opq(bbox = NULL)
}
\arguments{
\item{bbox}{base bounding box to use with the features. Must set the individual
feature bbox values if this value is not set.}
}
\value{
\code{opq} object
}
\description{
Begin building an Overpass query
}
\examples{
opq(bbox="43.0135509,-70.8229993,43.0996118,-70.7280563") \%>\%
  add_feature("amenity", "pub", ) \%>\%
  add_feature("amenity", "restaurant") \%>\%
  add_feature("amenity", "library") \%>\%
  issue_query() -> reading_noms

plot(reading_noms)
}

