% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opq.R
\name{opq_string}
\alias{opq_string}
\alias{opq_to_string}
\title{Convert an overpass query into a text string}
\usage{
opq_string(opq)
}
\arguments{
\item{opq}{An \code{overpass_query} object}
}
\value{
Character string to be submitted to the overpass API
}
\description{
Convert an osmdata query of class opq to a character string query to
be submitted to the overpass API.
}
\examples{
\dontrun{
q <- opq ("hampi india")
opq_string (q)
}
}
