% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{osm_elevation}
\alias{osm_elevation}
\title{osm_elevation}
\usage{
osm_elevation(dat, elev_file)
}
\arguments{
\item{dat}{An \code{SC} object produced by \link{osmdata_sc}.}

\item{elev_file}{A vector of one or more character strings specifying paths to
\code{.tif} files containing global elevation data.}
}
\value{
A modified version of the input \code{dat} with an additional \code{z_} column
appended to the vertices.
}
\description{
Add elevation data to a previously-extracted OSM data set, using a
pre-downloaded global elevation file from
\url{http://srtm.csi.cgiar.org/srtmdata/}. Currently only works for \code{SC}-class
objects returned from \link{osmdata_sc}.
}
