% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata-class.R, R/osmdata-package.R
\docType{package}
\name{osmdata}
\alias{osmdata}
\alias{osmdata-package}
\title{Create osmdata object}
\usage{
osmdata(bbox, overpass_call, osm_points, osm_lines, osm_polygons, timestamp,
  ...)
}
\arguments{
\item{bbox}{base bounding box to use with the features. Must set the individual
feature bbox values if this value is not set. Can be a matrix (i.e. what
\code{sp::bbox} returns), a string with values ("left,bottom,top,right"),
a vector of length 4, or (if tmap is installed) a character place name.
If the vector is named, the names will be used,
otherwise, you should ensure the vector is in \code{c(top, left, bottom, right)}
order.}
}
\description{
Initialise an empty or data-containing object of class \code{osmdata}

The Overpass API (or OSM3S) is a read-only API that serves up custom
selected parts of the OSM map data. It acts as a database over the web:
the client sends a query to the API and gets back the data set that
corresponds to the query.\cr
\cr
Unlike the main API, which is optimized for editing, Overpass API is
optimized for data consumers that need a few elements within a glimpse
or up to roughly 100 million elements in some minutes, both selected by
search criteria like e.g. location, type of objects, tag properties,
proximity, or combinations of them. It acts as a database backend for
various services.\cr
\cr
Overpass API has a powerful query language (language guide, language
reference, an IDE) beyond XAPI, but also has a compatibility layer to
allow a smooth transition from XAPI.
}
\section{Functions}{

\tabular{ll}{
\code{add_feature}\tab Add feature to an overpass API query\cr
\code{available_features}\tab Obtain list all OpenStreetMap features\cr
\code{available_tags}\tab List all tags for a given OpenStreetMap feature\cr
\code{bbox_to_string}\tab Convert matrix or vector to a bbox string to pass
to overpass API\cr
\code{issue_query}\tab Finalize and issue an Overpass query\cr
\code{opq}\tab Begin building an Overpass query\cr
\code{overpass_query}\tab Issue OSM Overpass Query\cr
\code{overpass_status}\tab Retrieve status of the Overpass API\cr
\code{read_osm}\tab Read an XML OSM Overpass response from path\cr
}
}
\author{
Bob Rudis, Robin Lovelace, Maelle Salmon, Mark Padgham
}

