% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata-class.R
\name{osmdata}
\alias{osmdata}
\title{osmdata class def}
\usage{
osmdata(
  bbox = NULL,
  overpass_call = NULL,
  meta = NULL,
  osm_points = NULL,
  osm_lines = NULL,
  osm_polygons = NULL,
  osm_multilines = NULL,
  osm_multipolygons = NULL
)
}
\arguments{
\item{bbox}{bounding box}

\item{overpass_call}{overpass_call}

\item{meta}{metadata of overpass query, including timestamps and version
numbers}

\item{osm_points}{OSM nodes as \pkg{sf} Simple Features Collection of points
or \pkg{sp} SpatialPointsDataFrame}

\item{osm_lines}{OSM ways \pkg{sf} Simple Features Collection of linestrings
or \pkg{sp} SpatialLinesDataFrame}

\item{osm_polygons}{OSM ways as \pkg{sf} Simple Features Collection of
polygons or \pkg{sp} SpatialPolygonsDataFrame}

\item{osm_multilines}{OSM relations as \pkg{sf} Simple Features Collection
of multilinestrings or \pkg{sp} SpatialLinesDataFrame}

\item{osm_multipolygons}{OSM relations as \pkg{sf} Simple Features
Collection of multipolygons or \pkg{sp}
SpatialPolygonsDataFrame}
}
\description{
osmdata class def
}
\note{
Class constructor should never be used directly, and is only exported
to provide access to the print method
}
\examples{
# This function should not need to be called directly!
osmdata ()
}
\concept{class}
