% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overpass_query.R
\name{overpass_query}
\alias{overpass_query}
\title{Issue OSM Overpass Query}
\usage{
overpass_query(query, quiet = FALSE, wait = TRUE, pad_wait = 5)
}
\arguments{
\item{query}{OSM Overpass query. Please note that the function is in ALPHA
dev stage and needs YOU to specify that the output type is XML.
However, you can use Overpass XML or Overpass QL formats.}

\item{quiet}{suppress status messages. OSM Overpass queries may not return
quickly. The package will display status messages by default showing
when the query started/completed.  You can disable these messages by
setting this value to \code{TRUE}.}

\item{wait}{if \code{TRUE} and if there is a queue at the Overpass API
server, should this function wait and try again at the next available
slot time or should it throw a an exception?}

\item{pad_wait}{if there is a queue and \code{wait} is \code{TRUE}, pad the
next query start time by \code{pad_wait} seconds (default = 5 seconds).}
}
\value{
If the \code{query} result only has OSM \code{node}s then the
        function will return a \code{SpatialPointsDataFrame} with the
        \code{node}s.\cr\cr
        If the \code{query} result has OSM \code{way}s then the function will
        return a \code{SpatialLinesDataFrame} with the \code{way}s\cr\cr
        \code{relations}s are not handled yet.\cr\cr
        If you asked for a CSV, you will receive the text response back,
        suitable for processing by \code{read.table(text=..., sep=...,
        header=TRUE, check.names=FALSE, stringsAsFactors=FALSE)}.
}
\description{
Issue OSM Overpass Query
}
\note{
wrap function with \code{httr::with_verbose} if you want to see the
      \code{httr} query (useful for debugging connection issues).\cr \cr 
      You can disable progress bars by calling
      \code{pbapply::pboptions(type="none")} in your code. See
      \code{\link[pbapply]{pboptions}} for all the various progress bar
      settings.
}
\examples{
\dontrun{
only_nodes <- '[out:xml];
node
  ["highway"="bus_stop"]
  ["shelter"]
  ["shelter"!~"no"]
  (50.7,7.1,50.8,7.25);
out body;'

pts <- overpass_query(only_nodes)
}
}

