% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_overpass_url}
\alias{set_overpass_url}
\title{set_overpass_url}
\usage{
set_overpass_url(overpass_url)
}
\arguments{
\item{overpass_url}{The desired overpass API URL}
}
\value{
The overpass API URL
}
\description{
Set the URL of the specified overpass API. Possible APIs with global coverage
are:
\itemize{
\item 'http://overpass-api.de/api/interpreter' (default)
\item 'http://overpass.osm.rambler.ru/cgi/interpreter'
\item 'http://api.openstreetmap.fr/oapi/interpreter'
\item 'https://overpass.osm.vi-di.fr/api/interpreter'
}
Additional APIs with limited local coverage include:
\itemize{
\item 'http://overpass.osm.ch/api/interpreter' (Switzerland)
\item 'http://overpass.openstreetmap.ie/api/interpreter' (Ireland)
}
}
\details{
For further details, see
\url{http://wiki.openstreetmap.org/wiki/Overpass_API}
}
\seealso{
\code{\link{get_overpass_url}}
}
