% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique-osmdata.R
\name{unique_osmdata}
\alias{unique_osmdata}
\title{unique_osmdata}
\usage{
unique_osmdata(dat)
}
\arguments{
\item{dat}{An \code{osmdata} object}
}
\value{
Equivalent object reduced to only unique objects of each type
}
\description{
Reduce the components of an \code{osmdata} object to only unique items of
each type. That is, reduce \code{$osm_points} to only those points not
present in other objects (lines, polygons, etc.); reduce \code{$osm_lines} to
only those lines not present in multiline objects; and reduce
\code{$osm_polygons} to only those polygons not present in multipolygon
objects. This renders an \code{osmdata} object more directly compatible with
typical output of \code{sf}.
}
